rename fpop to abacus
